Remove old vars, refactor tests (combine)
