+++ README.md
@@ -17,8 +17,6 @@ Defaults: `defaults/main.yml`
 
 - `postgresql_version`: The PostgreSQL major version: `9.6`, `10`, `11`
 - `postgresql_package_version`: The PostgreSQL full version, leave this empty to use the latest minor release from `postgresql_version`, ignored on Ubuntu
-- `postgresql_install_server`: If True (default) install and initialise the server (unless already installed and initialised), otherwise only install the client
-- `postgresql_install_extensions`: If `True` install extension (contrib) package, default `False`
 
 The following parameters will be ignored if `postgresql_install_server: False`:
 - `postgresql_databases`: List of dictionaries of databases.
