+++ defaults/main.yml
@@ -7,12 +7,6 @@
 # Full package version of postgres, see ome.postgresql_client
 # postgresql_package_version
 
-# Whether to install and setup PostgreSQL server
-postgresql_install_server: true
-
-# Install contrib extensions?
-postgresql_install_extensions: false
-
 # List of dictionaries of databases
 postgresql_databases: []
