+++ molecule/ubuntu1804/molecule.yml
@@ -8,6 +8,12 @@ driver:
 lint:
   name: yamllint
 platforms:
+  - name: postgresql-96-localhost
+    image: leandelivery/docker-systemd:ubuntu-18.04
+    command: /sbin/init
+    privileged: true
+    tmpfs:
+      - /sys/fs/cgroup
   - name: postgresql-10-localhost
     image: leandelivery/docker-systemd:ubuntu-18.04
     command: /sbin/init
@@ -15,7 +21,6 @@ platforms:
     tmpfs:
       - /sys/fs/cgroup
     groups:
-      - server
       - extra_options
   - name: postgresql-11-localhost
     image: leandelivery/docker-systemd:ubuntu-18.04
@@ -23,14 +28,14 @@ platforms:
     privileged: true
     tmpfs:
       - /sys/fs/cgroup
-    groups:
-      - server
 provisioner:
   name: ansible
   lint:
     name: ansible-lint
   inventory:
     host_vars:
+      postgresql-96-localhost:
+        postgresql_version: "9.6"
       postgresql-10-localhost:
         postgresql_version: "10"
         ansible_python_interpreter: /usr/bin/python3
@@ -39,7 +44,6 @@ provisioner:
         ansible_python_interpreter: /usr/bin/python3
     group_vars:
       extra_options:
-        postgresql_install_extensions: true
         postgresql_server_conf:
           shared_preload_libraries: "'pg_stat_statements'"
           log_filename: "'postgresql-%F.log'"
