+++ tasks/redhat.yml
@@ -14,7 +14,6 @@
       {{ postgresql_dist.redhat.basename }}-server{{
          postgresql_dist.redhat.version_suffix }}
     state: present
-  when: postgresql_install_server
 
 - name: postgres | install extension packages
   become: true
@@ -23,7 +22,6 @@
       {{ postgresql_dist.redhat.basename }}-contrib{{
          postgresql_dist.redhat.version_suffix }}
     state: present
-  when: postgresql_install_extensions
 
 - name: postgres | install ansible prerequisites
   become: true
@@ -36,4 +34,3 @@
     state: present
     disablerepo: "*"
     enablerepo: "{{ postgresql_dist.redhat.repoid }}"
-  when: postgresql_install_server
